* gpaste-3.30:
  gnome-shell: workaround a bug in gnome-shell 3.30.0
  Release GPaste 3.30.0
  update to gnome-shell 3.30
  bring other dependencies up to date
  use glib 2.58
  use gtk 3.24
  keybinder: silence gcc 8 warning

+++ src/gnome-shell/item.js
@@ -105,6 +105,10 @@ var GPasteItem = new Lang.Class({
 
     _onKeyPressed: function(actor, event) {
         const symbol = event.get_key_symbol();
+        if (symbol == Clutter.KEY_space || symbol == Clutter.KEY_Return) {
+            this.activate(event);
+            return Clutter.EVENT_STOP;
+        }
         if (symbol == Clutter.KEY_BackSpace || symbol == Clutter.KEY_Delete) {
             this._client.delete(this._index, null);
             return Clutter.EVENT_STOP;
